# docs/api/swagger.yaml
basePath: /api
definitions:
...
  internal_modules_rbac.AuthGetPermissionsRequest:
    properties:
      is_tree:
        type: boolean
    type: object
...
  internal_modules_rbac.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
  internal_modules_rbac.RbacApiCreateRequest:
    properties:
      name:
        maxLength: 20
        minLength: 2
        type: string
      parent_id:
        type: integer
      sort:
        type: integer
      url:
        type: string
    required:
    - name
    - url
    type: object
  internal_modules_rbac.RbacApiDeleteRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  internal_modules_rbac.RbacApiUpdateRequest:
    properties:
      id:
        type: integer
      name:
        maxLength: 20
        minLength: 2
        type: string
      parent_id:
        type: integer
      sort:
        type: integer
      url:
        type: string
    required:
    - id
    - name
    - url
    type: object
  internal_modules_rbac.RbacMenu:
    properties:
      action_mark:
        type: string
      actions:
        items:
          $ref: '#/definitions/internal_modules_rbac.RbacMenu'
        type: array
      children:
        items:
          $ref: '#/definitions/internal_modules_rbac.RbacMenu'
        type: array
      created_at:
        type: integer
      id:
        type: integer
      is_page:
        type: integer
      module_key:
        type: string
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      sort:
        type: integer
      type:
        type: integer
      updated_at:
        type: integer
    type: object
  internal_modules_rbac.RbacMenuApiListRequest:
    properties:
      menu_id:
        type: integer
    required:
    - menu_id
    type: object
  internal_modules_rbac.RbacMenuApiSaveRequest:
    properties:
      api_ids:
        items:
          type: integer
        type: array
      menu_id:
        type: integer
    required:
    - menu_id
    type: object
  internal_modules_rbac.RbacMenuCreateRequest:
    properties:
      action_mark:
        type: string
      is_page:
        type: integer
      module_key:
        type: string
      name:
        maxLength: 20
        minLength: 2
        type: string
      parent_id:
        type: integer
      path:
        type: string
      sort:
        type: integer
      type:
        type: integer
    required:
    - name
    - path
    - type
    type: object
  internal_modules_rbac.RbacMenuDeleteRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  internal_modules_rbac.RbacRoleCreateRequest:
    properties:
      parent_id:
        type: integer
      role_name:
        maxLength: 50
        minLength: 2
        type: string
      sort:
        type: integer
      store_id:
        type: integer
    required:
    - role_name
    type: object
...
  internal_modules_user.UserLoginResponse:
    properties:
      token:
        type: string
      ttl:
        type: integer
      user:
        $ref: '#/definitions/internal_modules_user.User'
    type: object
  response.Response:
    properties:
      cost:
        type: string
      data: {}
      errcode:
        type: integer
      message:
        type: string
      traceId:
        type: string
    type: object
  zero-backend_internal_modules_rbac.AuthLoginRequest:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha_code
    - captcha_id
    - password
    - username
    type: object
  zero-backend_internal_modules_rbac.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 20
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
host: localhost:8081
info:
  contact: {}
  title: Zero API
  version: "1.0"
paths:
...
  /change-password:
    post:
      parameters:
      - description: 修改密码参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/zero-backend_internal_modules_rbac.ChangePasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户修改密码
      tags:
      - 用户认证
...
  /login:
    post:
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/zero-backend_internal_modules_rbac.AuthLoginRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_user.UserLoginResponse'
              type: object
      summary: 用户登录
      tags:
      - 用户认证
...
  /permissions:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.AuthGetPermissionsRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_modules_rbac.RbacMenu'
                  type: array
              type: object
      summary: 获取用户权限菜单
      tags:
      - 认证
  /rbac/api/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacApiCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建接口
      tags:
      - RBAC接口管理
  /rbac/api/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacApiDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除接口
      tags:
      - RBAC接口管理
...
  /rbac/api/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacApiUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新接口
      tags:
      - RBAC接口管理
  /rbac/menu/api/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacMenuApiListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_rbac.ListResult'
              type: object
      summary: 获取菜单关联的接口列表
      tags:
      - RBAC菜单管理
  /rbac/menu/api/save:
    post:
      parameters:
      - description: 保存参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacMenuApiSaveRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 保存菜单关联的接口列表
      tags:
      - RBAC菜单管理
  /rbac/menu/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacMenuCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建菜单
      tags:
      - RBAC菜单管理
  /rbac/menu/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacMenuDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除菜单
      tags:
      - RBAC菜单管理
...
  /rbac/menu/sync:
    post:
      parameters:
      - description: 同步参数
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/internal_modules_rbac.RbacMenuSyncRequest'
          type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 同步菜单
      tags:
      - RBAC菜单管理
  /rbac/menu/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacMenuUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新菜单
      tags:
      - RBAC菜单管理
  /rbac/role/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacRoleCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建角色
      tags:
      - RBAC角色管理
  /rbac/role/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacRoleDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除角色
      tags:
      - RBAC角色管理
  /rbac/role/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacRoleListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_rbac.ListResult'
              type: object
      summary: 获取角色列表
      tags:
      - RBAC角色管理
  /rbac/role/set-menus:
    post:
      parameters:
      - description: 设置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacRoleMenuSetRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 设置角色菜单
      tags:
      - RBAC角色管理
  /rbac/role/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacRoleUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新角色
      tags:
      - RBAC角色管理
  /rbac/store/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacStoreCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建企业
      tags:
      - RBAC企业管理
  /rbac/store/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacStoreDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除企业
      tags:
      - RBAC企业管理
  /rbac/store/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacStoreListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_rbac.ListResult'
              type: object
      summary: 获取企业列表
      tags:
      - RBAC企业管理
  /rbac/store/recycle:
    post:
      parameters:
      - description: 回收参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacStoreDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 回收企业
      tags:
      - RBAC企业管理
  /rbac/store/restore:
    post:
      parameters:
      - description: 恢复参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacStoreDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 恢复企业
      tags:
      - RBAC企业管理
  /rbac/store/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacStoreUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新企业
      tags:
      - RBAC企业管理
  /rbac/user/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacUserCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建后台用户
      tags:
      - RBAC用户管理
  /rbac/user/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacUserDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除后台用户
      tags:
      - RBAC用户管理
  /rbac/user/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacUserListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_rbac.ListResult'
              type: object
      summary: 获取后台用户列表
      tags:
      - RBAC用户管理
  /rbac/user/reset-password:
    post:
      parameters:
      - description: 重置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacUserResetPasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 重置用户密码
      tags:
      - RBAC用户管理
  /rbac/user/set-roles:
    post:
      parameters:
      - description: 设置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacUserRoleSetRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 设置用户角色
      tags:
      - RBAC用户管理
  /rbac/user/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_rbac.RbacUserUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新后台用户
      tags:
      - RBAC用户管理
...
swagger: "2.0"